+++ taskflow/src/main/java/cn/taskflow/conductor/core/secrets/JdbcSecretStore.java
@@ -32,8 +32,8 @@ public final class JdbcSecretStore implements SecretStore {
             CREATE TABLE IF NOT EXISTS taskflow_secret (
               name VARCHAR(100) NOT NULL PRIMARY KEY,
               value TEXT NOT NULL,
-              created_at INTEGER NOT NULL,
-              updated_at INTEGER NOT NULL
+              created_at BIGINT NOT NULL,
+              updated_at BIGINT NOT NULL
             )
             """;
